refactor: rename corotuines related package in core
