+++ core/src/main/java/com/monstarlab/core/coroutines/CoroutinesModule.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.core.injection
+package com.monstarlab.core.coroutines
 
 import dagger.Module
 import dagger.Provides
